Agent compiles on macOS
