+++ changelog.d/2453.internal.md
@@ -1,0 +1,1 @@
+Agent compiles on macOS
